+++ DockDoor/Localizable.xcstrings
@@ -9353,9 +9353,6 @@
         }
       }
     },
-    "Button Position" : {
-
-    },
     "Buy me a coffee here, thank you!" : {
       "localizations" : {
         "af" : {
