+++ DockDoor/Views/Settings/Help/BaseSettingsView.swift
@@ -10,10 +10,8 @@ struct BaseSettingsView<Content: View>: View {
     var body: some View {
         ScrollView {
             content
-                .padding(20) // Apply padding to the content itself inside the ScrollView
+                .padding(20)
         }
-        // This frame makes the content area of each tab uniform.
-        // It allows the width to expand if the window is wider, but fixes the max height.
         .frame(minWidth: 650, idealWidth: 700, maxWidth: .infinity,
                minHeight: 675, idealHeight: 700, maxHeight: 700)
     }
